Adding labels to drivers
This change adds labels to the GKE and AKS drivers. When a cluster is
created, a label or tag will be added to the GKE, AKS or EKS cluster
respectively.  The value of the value of the label or tag will be the
display name of the cluster in the rancher UI, allowing the end-user to
better identify what resources correspond to what cluster in Rancher.

Issue:
https://github.com/rancher/rancher/issues/12533