Fixing EKS credentials issue
This change fixes the EKS credentials issue where it was not able to
generate a bearer token to apply the config map in the EKS cluster after
it was created.  It fixes this issue by writing a configuration file just
before the heptio sdk is used to generate the bearer token, and then
cleaning it up immediately after.

Issue:
https://github.com/rancher/rancher/issues/12589