Fixing EKS driver panic
Fixing an issue where a panic from the EKS driver would bring down
rancher.  When a cluster is not found, the cluster struct is instantiated
without a status field.  When the code tries to dereference the status
pointer a panic brings down rancher. This change introduces a nil check
before dereferncing the pointer. It also adds a status code check to the
aws http requests so that a failed cluster create would be caught earlier
in the code.

Issue:
https://github.com/rancher/rancher/issues/13091