Send EnableRbac=true on cluster create, add logs
AKS by default will set EnabledRBAC to false, so it doesn't create any built in roles. When we filter projects/workloads etc for a cluster, we don't get back any role info (since the default roles we map to do not exist), and nothing is shown. This sets EnableRBAC to true when we create a cluster, but it omits it when we update, since the value is not allowed to be changed (or the update will error out).

It also logs the config we send to azure (debug level log) with the clientId/secret redacted.