Adding ability to pass in custom ami
This change adds the ability to pass in a custom ami to EKS driver.  This is
to allow users the ability to use their own custom images instead of the EKS
default ones.  Also some logic was added to allow new regions to be passed in
without kontainer-engine being updated to accomodate them, As long as an AMI
ID is given, EKS driver will not fail for a unknown region during validation.

Issue:
https://github.com/rancher/rancher/issues/15389