Changing EKS driver to use display name instead of cluster name
This change makes it so that resources created in AWS during EKS cluster
creation are prepended with the display name of the cluster in rancher.

This also adds backwards support for clusters with old style names so that
they can still be deleted successfully (update is not supported in EKS
currently).

Issue:
https://github.com/rancher/rancher/issues/15390