Adding AssociatePublicIpAddress to EKS driver
Adding an option to EKS driver to set the value of AssociatePublicIPAddress
in the worker nodes template.  This is so users can create a private cluster
with rancher with properly configured vpc and subnets.

Also adds validation to prevent AssociatePublicIPAddress from being set to
`false` without also passing in a custom vpc and subnets.

Issue:
https://github.com/rancher/rancher/issues/15419