Adding code to properly lookup alternate eks cluster names
This change adds logic to properly look up alternate (legacy) naming
convention of EKS clusters when performing an update or generating the
clientset to communicate with the k8s api.  This is so that legacy EKS
clusters will continue to work after upgrading to the latest version of
rancher.

Issue:
https://github.com/rancher/rancher/issues/18162