gke: Refactor Projects.Zonal to Projects.Locations
The GKE SDK has 2 services, Projects.Zonal and Projects.Locations.
Project.Zonals locks you into only being able to CRUD on zonal
resources. Projects.Locations takes Relative Resource Names instead
letting you CRUD both zonal and regional resources. This is made
possible by the fact that the RRNs contain the location which can
be eiter a zone or a region.

This commit switches to using Projects.Locations in order to be able to
support regional clusters later on. It also introduces 3 helper
functions to correctly generate the RRNs that we now need to pass to
most of the methods that get called on Projects.Locations.

We also stop setting the deprecated Cluster.Zone attribute. At
creation time we pass a RRN representing a location to
Projects.Locations.Create() which it will use to decide whether to
create a zonal or regional cluster.