gke: Enable regional clusters
This commit enables the creation of regional clusters. We do this by
adding a Region parameter to the state object and validating that only
one of region or zone is set. You cannot set both Region and Zone so we
have to drop the default for Zone otherwise we'll always get caught by
the check in validate().

We add a location() helper to state which returns either the Region or
Zone the state declares. This is then used in place where we'd
normally use state.Zone. Based on whether Region or Zone is set we're
now generating RRNs with a region or zone in the location, letting us
act on either regional or zonal clusters.