Update EKS default k8s version
Updated EKS default k8s version for create/update to 1.13. Prior,
it was 1.10. 1.10 is no longer supported for EKS and keeping it as
the value for k8s version parameter would cause an error. Now, the
default k8s version for EKS is the same as the on in the AWS
console.