Do not error on AKS upgrading status
AKS driver will no longer error out when observing a cluster
status of "Upgrading" from AKS API response. Prior, when the
Kubernetes version of an AKS cluster was upgraded, AKS would
respond with a cluster status of "Upgrading". Rancher evaluated
this as an error, which was incorrect and caused a failed
provisioning error to propagate to the UI. Rancher would then
retry an update call to the AKS API which resulted in a
legitimate error, because the cluster was already being
upgraded. AKS/EKS/GKE have also had some of their logs modified
to include the name of the driver that produced them and any
relevant cluster IDs.